Removed an unecessary component yay
